[MODORDERS-501] - Move InventoryManager to services
