+++ src/test/java/org/folio/orders/events/handlers/CheckInOrderStatusChangeChangeHandlerTest.java
@@ -14,7 +14,7 @@ import static org.folio.TestConstants.PO_ID_PENDING_STATUS_WITH_PO_LINES;
 import static org.folio.TestUtils.checkVertxContextCompletion;
 import static org.folio.helper.AbstractHelper.ORDER_ID;
 import static org.folio.helper.CheckinHelper.IS_ITEM_ORDER_CLOSED_PRESENT;
-import static org.folio.helper.InventoryManager.ITEMS;
+import static org.folio.service.inventory.InventoryManager.ITEMS;
 import static org.folio.rest.impl.MockServer.ITEM_RECORDS;
 import static org.folio.rest.impl.MockServer.getItemUpdates;
 import static org.folio.rest.impl.MockServer.getItemsSearches;
@@ -47,10 +47,8 @@ import org.folio.ApiTestSuite;
 import org.folio.config.ApplicationConfig;
 import org.folio.helper.CheckinHelper;
 import org.folio.orders.utils.HelperUtils;
-import org.folio.rest.core.RestClient;
 import org.folio.rest.jaxrs.model.PurchaseOrder;
 import org.folio.rest.jaxrs.model.PurchaseOrder.WorkflowStatus;
-import org.folio.service.configuration.ConfigurationEntriesService;
 import org.folio.service.finance.transaction.EncumbranceService;
 import org.folio.service.orders.PurchaseOrderService;
 import org.folio.spring.SpringContextUtil;
@@ -61,7 +59,6 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
 
 import io.vertx.core.AsyncResult;
 import io.vertx.core.Handler;
