+++ src/test/java/org/folio/rest/impl/MockServer.java
@@ -39,9 +39,9 @@ import static org.folio.TestConstants.PROTECTED_READ_ONLY_TENANT;
 import static org.folio.TestConstants.X_ECHO_STATUS;
 import static org.folio.TestUtils.getMockAsJson;
 import static org.folio.TestUtils.getMockData;
-import static org.folio.helper.InventoryManager.ITEMS;
-import static org.folio.helper.InventoryManager.ITEM_PURCHASE_ORDER_LINE_IDENTIFIER;
-import static org.folio.helper.InventoryManager.REQUESTS;
+import static org.folio.service.inventory.InventoryManager.ITEMS;
+import static org.folio.service.inventory.InventoryManager.ITEM_PURCHASE_ORDER_LINE_IDENTIFIER;
+import static org.folio.service.inventory.InventoryManager.REQUESTS;
 import static org.folio.helper.InventoryManagerTest.HOLDING_INSTANCE_ID_2_HOLDING;
 import static org.folio.helper.InventoryManagerTest.NEW_LOCATION_ID;
 import static org.folio.helper.InventoryManagerTest.NON_EXISTED_NEW_HOLDING_ID;
